+++ src/main/resources/db/migration/V02__criar_e_registrar_cidades.sql
@@ -1,6 +1,6 @@
 CREATE TABLE TB_CIDADE
 (
-	cod 		INT 		AUTO_INCREMENT,
+	cod 		BIGINT 		AUTO_INCREMENT,
     nome_cidade VARCHAR(100) NOT NULL,
     CONSTRAINT pk_tb_cidade 		PRIMARY KEY(cod),
     CONSTRAINT uq_tb_cidade_nome 	UNIQUE(nome_cidade)
