+++ src/main/resources/db/migration/V4__criar_e_registrar_tipos_de_telefone.sql
@@ -1,6 +1,6 @@
 CREATE TABLE TB_TIPO_TELEFONE
 (
-	cod 	INT 		AUTO_INCREMENT,
+	cod 	BIGINT 		AUTO_INCREMENT,
     tipo 	VARCHAR(20) NOT NULL,
     CONSTRAINT pk_tb_tipo_telefone PRIMARY KEY(cod),
     CONSTRAINT uq_tb_tipo_telefone UNIQUE(tipo)
